requirements.txt
# These are the required modules to run the programs within this module
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.5
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.17.0
